Add utility: creator for magic pixels (#3)
Add a utility class `RotmanMegProjector` to wrap around the Propixx projector in the MEG lab.
Class contains site-specific parameters that are stable and ready to use.
Enables end users to call helper functions to generate a magic pixel square on top-left corner of the screen.
Enables end users to get a handle to the `window` object.
Enables end users to convert seconds into number of frames.